feat/playground: Создано демо
Добавлено демо с возможностью перезагрузки при редактировании файлов библиотеки
