+++ .gitignore
@@ -9,6 +9,7 @@
 
 # Node
 /node_modules
+/projects/demo/node_modules
 npm-debug.log
 yarn-error.log
 package-lock.json
@@ -42,3 +43,4 @@ testem.log
 # System files
 .DS_Store
 Thumbs.db
+/projects/demo/node_modules/
