+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@angular/animations": "^18.2.0",
     "@angular/common": "^18.2.0",
-    "@angular/compiler": "^18.2.0",
+    "@angular/compiler": "^18.2.13",
     "@angular/core": "^18.2.0",
     "@angular/forms": "^18.2.0",
     "@angular/platform-browser": "^18.2.0",
