+++ projects/demo/ng-icons/ng-icons-list.ts
@@ -1,0 +1,13 @@
+import {
+  faSolidCircleXmark,
+  faSolidEye,
+  faSolidEyeSlash,
+  faSolidXmark
+} from '@ng-icons/font-awesome/solid';
+
+export const ngIconsList = {
+  faSolidXmark,
+  faSolidCircleXmark,
+  faSolidEye,
+  faSolidEyeSlash
+}
