+++ projects/demo/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "@angular/compiler": "^18.2.0"
+  }
+}
