+++ projects/demo/src/app/app.component.ts
@@ -2,6 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
-  template: '<router-outlet />',
+  template: '<router-outlet />'
 })
 export class AppComponent {}
