+++ projects/demo/src/app/app.module.ts
@@ -3,6 +3,9 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import {UiKitModule} from '@egor-bamburov/ng-ui-kit';
+import {NgIconsModule} from '@ng-icons/core';
+import {ngIconsList} from '../../../../src/app/ng-icons/ng-icons-list';
 
 @NgModule({
   declarations: [
@@ -11,6 +14,8 @@ import { AppComponent } from './app.component';
   imports: [
     BrowserModule,
     AppRoutingModule,
+    NgIconsModule.withIcons(ngIconsList),
+    UiKitModule
   ],
   providers: [],
   bootstrap: [AppComponent]
