+++ projects/demo/src/app/pages/inputs/inputs.page.html
@@ -1,0 +1,2 @@
+<ui-kit-text-input placeholder="test" label="test" errorLabel="test" />
+<ui-kit-password-input placeholder="test" />
