+++ projects/demo/src/app/pages/inputs/inputs.page.ts
@@ -1,17 +1,14 @@
 import { Component } from '@angular/core';
-// import {UiKitModule} from '../../../../projects/ng-ui-kit/src/public-api'
+import {UiKitModule} from '@egor-bamburov/ng-ui-kit';
 
 @Component({
   selector: 'app-inputs',
   standalone: true,
   imports: [
-    // UiKitModule
+    UiKitModule
   ],
   templateUrl: './inputs.page.html',
   styleUrl: './inputs.page.scss'
 })
 export class InputsPage {
-constructor() {
-  console.log('123')
-}
 }
