+++ projects/demo/src/app/pages/toasts/toasts.page.html
@@ -1,0 +1,1 @@
+<ui-kit-toast defaultDuration="1000" />
