+++ projects/demo/src/styles.scss
@@ -1,1 +1,2 @@
 /* You can add global styles to this file, and also import other style files */
+@import "@egor-bamburov/ng-ui-kit/styles/styles";
