--- src/app/app.component.scss
@@ -1,1 +1,0 @@
-
