+++ src/app/app.routes.ts
@@ -1,16 +1,3 @@
 import {Routes} from '@angular/router';
-import {MainRoutes} from './layouts/main/main.routes';
 
-export const AppRoutes: Routes = [
-  {
-    path: '',
-    loadComponent: () => import('./layouts/main/main.layout')
-      .then((l) => l.MainLayout),
-    children: MainRoutes
-  },
-  {
-    path: '**',
-    loadComponent: () => import('./pages/not-found/not-found.page')
-      .then((p) => p.NotFoundPage)
-  }
-]
+export const AppRoutes: Routes = []
