--- src/app/pages/inputs/inputs.page.html
@@ -1,1 +1,0 @@
-<!--<ui-kit-text-input placeholder="test" label="test" errorLabel="test" />-->
