--- src/app/pages/toasts/toasts.page.html
@@ -1,1 +1,0 @@
-<!--<ui-kit-toast defaultDuration="1000" />-->
